break-up test file and enable GV for merging users in tests
